edit: deleted searchconsultant folder now it is in booking.also ma index that redirects to searchconsultant component. rename: renamed some folders in the components dir still need to switch a couple more
