+++ app/(booking)/_layout.js
@@ -3,6 +3,7 @@ import { Stack } from 'expo-router';
 import { theme } from '../../styles/theme';
 import BookingProvider from '../../context/BookingContext';
 import { ConsultantFiltersProvider } from '../../context/ConsultantFilterContext';
+
 const Layout = () => {
 	return (
 		<ConsultantFiltersProvider>
