+++ components/searching-consultant/consultantCard/ConsultantCardList.js
@@ -5,14 +5,14 @@ import { useFonts } from 'expo-font';
 import { fonts } from '../../../styles/fonts';
 import Loading from '../../common/Loading';
 import ErrorMessage from '../common/ErrorView';
-import ConsultantItem from './ConsultantItem';
+import ConsultantCardItem from './ConsultantCardItem';
 
 import errorMessages from '../../../locales/en/ErrorMessages.json';
 //This is the consultant card component that renders one of 3 things.
 //1.An error message if there is no consultants that match that criteria,
 //2.another error message that is when a fetch request fails.
 //3.he last return statement just returns the card that shows the consultant
-const ConsultantList = ({ consultantData, isLoading, errorMessage }) => {
+const ConsultantCardList = ({ consultantData, isLoading, errorMessage }) => {
 	const [fontsLoaded] = useFonts(fonts);
 
 	if (!fontsLoaded) {
@@ -53,7 +53,7 @@ const ConsultantList = ({ consultantData, isLoading, errorMessage }) => {
 	) : (
 		<View style={styles.container}>
 			{consultantData.map((consultant, index) => (
-				<ConsultantItem key={index} consultant={consultant} />
+				<ConsultantCardItem key={index} consultant={consultant} />
 			))}
 		</View>
 	);
@@ -92,4 +92,4 @@ const styles = StyleSheet.create({
 	},
 });
 
-export default ConsultantList;
+export default ConsultantCardList;
