+++ components/searching-consultant/filters/filterModalComponents/FilterModal.js
@@ -26,7 +26,7 @@ const FilterModal = ({ modalVisible, closeModal, filterContent }) => {
 	);
 };
 
-import { theme } from '../../../styles/theme';
+import { theme } from '../../../../styles/theme';
 
 const styles = StyleSheet.create({
 	modalContainer: {
