+++ components/searching-consultant/tabs/FilterTabs.js
@@ -4,7 +4,7 @@ import { useFonts } from 'expo-font';
 
 import { fonts } from '../../../styles/fonts';
 import FilterIcon from '../../../assets/svg/FilterIcon';
-import FilterModal from '../filters/FilterModal';
+import FilterModal from '../filters/filterModalComponents/FilterModal';
 import filteringOptions from '../../../defaultdata/Filtering-Tabs.json';
 import Button from '../../common/Button';
